Increase strictness.
Also delineates slightly better beween flexibility for array keys
(in Cast & Iter) and lack of it for array values.

Split TypeIs::listString() from TypeIs::stringOrListString() and made
it not use Cast, because TypeIs is supposed to be unforgiving. (I.e.,
you used TypeIs, you're telling me this is already true.)